shared types between front & backend
