+++ app/src/types/event.ts
@@ -2,27 +2,31 @@ export interface EventPayload {
     id: string;
     title: string;
     description: string;
-    location: string;
-    startDate: number;
-    endDate: number;
-    reviewers: string[];
-    managers: string[];
+    userId: string;
     challenges: string[];
+    version: number;
+    isNew: boolean;
+    createdAt: number;
+    updatedAt: number;
+    startDate?: number;
+    endDate?: number;
+    reviewers?: string[];
+    managers?: string[];
 }
 
 export interface CreateEventPayload {
+    id: string;
     title: string;
     description: string;
 }
 
 export interface UpdateEventPayload {
-    title: string;
-    description: string;
-    startDate: number;
-    endDate: number;
-    challenges: string[];
-    reviewers: string[];
-    managers: string[];
+    id: string;
+    data: {
+        title: string;
+        description: string;
+        challenges: string[];
+    };
 }
 
 export interface EventSettingsFormData {
