+++ app/src/types/leader-board.ts
@@ -7,3 +7,7 @@ export interface LeaderBoardPayload {
     totalPoints: number;
     participants: ParticipantPayload[];
 }
+
+export interface UpdateLeaderBoardPayload {
+    eventId: string;
+}
