+++ app/src/types/user.ts
@@ -9,6 +9,12 @@ export type UserPayload = {
     level?: number;
 };
 
+export interface SetUserPayload {
+    fullName: string;
+    userName: string;
+    walletPublicKey: string;
+}
+
 export interface UpdateUserFormData {
     fullName: string;
     userName: string;
