+++ app/src/utils/firebase.ts
@@ -10,7 +10,7 @@ import {
 } from 'firebase/auth';
 import { connectFirestoreEmulator, getFirestore } from 'firebase/firestore';
 import { connectFunctionsEmulator, getFunctions } from 'firebase/functions';
-import { AuthProviderType } from 'types/api';
+import { AuthProviderType } from 'types';
 
 const firebaseConfig = {
     apiKey: process.env.NEXT_PUBLIC_API_KEY,
