+++ backend/functions/package.json
@@ -13,7 +13,7 @@
     "engines": {
         "node": "16"
     },
-    "main": "lib/index.js",
+    "main": "lib/backend/functions/src/index.js",
     "dependencies": {
         "@solana/web3.js": "^1.66.2",
         "body-parser": "^1.20.1",
