+++ backend/functions/src/service/user.ts
@@ -1,6 +1,5 @@
 import * as functions from 'firebase-functions';
-import { db } from '..';
-import { Auth, SetUserPayload } from '../util/types';
+import { Auth, db, SetUserPayload } from '..';
 
 export const isDuplicateUserName = async (
     currentUserId: string,
@@ -17,7 +16,7 @@ export const isDuplicateUserName = async (
     );
 };
 
-class UserController {
+class UserService {
     async setUser(auth: Auth, payload: SetUserPayload) {
         if (await isDuplicateUserName(auth.id, payload.userName)) {
             throw new functions.https.HttpsError(
@@ -37,4 +36,4 @@ class UserController {
     }
 }
 
-export const controller = new UserController();
+export const userService = new UserService();
